requirements.txt
# HARBench Artifact Requirements
# PyTorch (install separately with CUDA support)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0

# Deep learning utilities
einops>=0.4.0
timm>=0.6.0

# Foundation models (optional)
transformers>=4.30.0
momentfm>=0.1.0

# Preprocessing (har-datasets)
pyyaml>=6.0
requests>=2.31.0

# Visualization (optional)
plotly>=5.17.0
flask>=3.0.0

# Development & Testing (optional)
pytest>=7.4.0
